Fix pow tests with RTools 3.5
